Added code review changes
Added the changes suggested in the code review.

Also removed the expected failure in LLVM 2.9. Will add if it will be
necessary.
