+++ tools/klee/main.cpp
@@ -191,8 +191,8 @@ namespace {
   SeedOutDir("seed-out-dir");
 
   cl::list<std::string>
-  LinkLibraries("link-libraries",
-		cl::desc("Link the give libraries before execution"),
+  LinkLibraries("link-llvm-lib",
+		cl::desc("Link the given libraries before execution"),
 		cl::value_desc("library file"));
 
   cl::opt<unsigned>
@@ -1308,13 +1308,13 @@ int main(int argc, char **argv, char **envp) {
     assert(mainModule && "unable to link with simple model");
   }
 
-  // TODO(omeranson) I think it should go here.
   std::vector<std::string>::iterator libs_it;
   std::vector<std::string>::iterator libs_ie;
   for (libs_it = LinkLibraries.begin(), libs_ie = LinkLibraries.end();
           libs_it != libs_ie; ++libs_it) {
-    llvm::errs() << "Linking in library: " << *libs_it << ".\n";
-    mainModule = klee::linkWithLibrary(mainModule, libs_it->c_str());
+    const char * libFilename = libs_it->c_str();
+    klee_message("Linking in library: %s.\n", libFilename);
+    mainModule = klee::linkWithLibrary(mainModule, libFilename);
   }
   // Get the desired main function.  klee_main initializes uClibc
   // locale and other data and then calls main.
